more work on nested-set
