+++ app/Http/Controllers/SearchController.php
@@ -5,8 +5,9 @@
 namespace App\Http\Controllers;
 
 use AccessControl;
+use App\Actions\Album\Cast as AlbumCast;
+use App\Actions\Albums\Top;
 use App\Actions\ReadAccessFunctions;
-use App\ModelFunctions\AlbumActions\Cast as AlbumCast;
 use App\ModelFunctions\AlbumFunctions;
 use App\ModelFunctions\AlbumsFunctions;
 use App\ModelFunctions\PhotoActions\Cast as PhotoCast;
@@ -32,11 +33,6 @@ class SearchController extends Controller
 	private $albumsFunctions;
 
 	/**
-	 * @var SessionFunctions
-	 */
-	private $sessionFunctions;
-
-	/**
 	 * @var readAccessFunctions
 	 */
 	private $readAccessFunctions;
@@ -45,6 +41,11 @@ class SearchController extends Controller
 	 * @var SymLinkFunctions
 	 */
 	private $symLinkFunctions;
+
+	/**
+	 * @var Top
+	 */
+	private $top;
 
 	/**
 	 * @param AlbumFunctions      $albumFunctions
@@ -57,12 +58,14 @@ class SearchController extends Controller
 		AlbumFunctions $albumFunctions,
 		AlbumsFunctions $albumsFunctions,
 		ReadAccessFunctions $readAccessFunctions,
-		SymLinkFunctions $symLinkFunctions
+		SymLinkFunctions $symLinkFunctions,
+		Top $top
 	) {
 		$this->albumFunctions = $albumFunctions;
 		$this->albumsFunctions = $albumsFunctions;
 		$this->readAccessFunctions = $readAccessFunctions;
 		$this->symLinkFunctions = $symLinkFunctions;
+		$this->top = $top;
 	}
 
 	/**
@@ -131,8 +134,7 @@ class SearchController extends Controller
 		 * from the top level.  This includes password-protected albums
 		 * (since they are visible) but not their content.
 		 */
-		$toplevel = $this->albumsFunctions->getToplevelAlbums();
-		$albumIDs = $this->albumsFunctions->getPublicAlbumsId($toplevel, null, true);
+		$albumIDs = $this->albumsFunctions->getPublicAlbumsId();
 
 		$query = Album::with([
 			'owner',
@@ -183,7 +185,7 @@ class SearchController extends Controller
 		 * accessible from the top level, only this time without
 		 * password-protected ones.
 		 */
-		$albumIDs = $this->albumsFunctions->getPublicAlbumsId($toplevel);
+		$albumIDs = $this->albumsFunctions->getPublicAlbumsId();
 		$query = Photo::with('album')->where(
 			function (Builder $query) use ($albumIDs) {
 				$query->whereIn('album_id', $albumIDs);
