+++ config/app.php
@@ -214,7 +214,7 @@ return [
 		'Mail' => Illuminate\Support\Facades\Mail::class,
 		'Markdown' => GrahamCampbell\Markdown\Facades\Markdown::class,
 		'Notification' => Illuminate\Support\Facades\Notification::class,
-		'Password' => Illuminate\Support\Facades\Password::class,
+		// 'Password' => Illuminate\Support\Facades\Password::class,
 		'Queue' => Illuminate\Support\Facades\Queue::class,
 		'Redirect' => Illuminate\Support\Facades\Redirect::class,
 		'Redis' => Illuminate\Support\Facades\Redis::class,
